Finish setting up user module and refactoring
