+++ Leaderboard/package.json
@@ -8,6 +8,7 @@
   },
   "dependencies": {
     "babel": "^5.8.23",
+    "bcryptjs": "^2.3.0",
     "body-parser": "~1.13.2",
     "config": "^1.16.0",
     "cookie-parser": "~1.3.5",
