+++ Leaderboard/views/layout.hbs
@@ -2,7 +2,7 @@
 <html>
   <head>
     <title>{{appName}}</title>
-    <link rel="stylesheet" href="https://storage.googleapis.com/code.getmdl.io/1.0.4/material.{{primaryColor}}-{{accentColor}}.min.css">
+    <link rel="stylesheet" href="https://storage.googleapis.com/code.getmdl.io/1.0.5/material.{{primaryColor}}-{{accentColor}}.min.css">
     <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
     <link rel='stylesheet' href='/stylesheets/style.css' />
   </head>
